feat: set up redux
