+++ src/action/type.js
@@ -1,0 +1,2 @@
+// Offer
+export const GET_OFFERS = 'get_offers';
