+++ src/api/get.js
@@ -1,3 +1,17 @@
 import data from './demo';
 
-export default data;
+class OfferAPI {
+    static async getPostList() {
+        try {
+            // const response = await fetchApi(
+            //     `whale.drkbl.com/getOffers`,
+            //     'GET'
+            // );
+            return await data; // return the dummy data for now
+        } catch (error) {
+            throw error;
+        }
+    }
+}
+
+export default OfferAPI;
