+++ src/component/offer/index.js
@@ -1,0 +1,3 @@
+import OfferList from './OfferList';
+
+export { OfferList };
