+++ src/reducer/index.js
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux';
+import OfferReducer from './OfferReducer';
+
+export default combineReducers({
+    offer: OfferReducer
+});
